refactor: move rendering logic to main banner component
